Bugfix: unlike for phonons, helper_el_scattering must always compute the bandstructure, or you may use the electronic energies. Bugfix: helper_el_scattering was using Points->getIndex with cartesian coordinates instead of crystal. Bugfix: helper_el_scattering was not distinguishing between activePoints and FullPoints when calling getIndex().

+++ src/bands/active_bandstructure.cpp
@@ -748,7 +748,7 @@ StatisticsSweep ActiveBandStructure::buildAsPostprocessing(
 
   // now we can loop over the trimmed list of points
   for (int i=0; i<filter.size(); i++ ) {
-    int ik = filter(i);
+    long ik = filter(i);
     auto ikIndex = WavevectorIndex(ik);
     Point point = activePoints.getPoint(i);
